Use EWA Weights instead of normal weights for weight loading. Switch to 1.1e-5 epsilon

+++ _weight_extraction_v2/extract_weights.py
@@ -65,8 +65,16 @@ inspect checkpoint --file_name = {}"""
 print_tensors_in_checkpoint_file(checkpoint_file, None, True)
 
 os.remove('weights/global_step.npy')
-for fn in glob.glob('weights/*_ExponentialMovingAverage.npy'):
-    os.remove(fn)
+
+# remove EMA Weights
+#for fn in glob.glob('weights/*_ExponentialMovingAverage.npy'):
+#   os.remove(fn)
+
+# Remove non EMA weights
+for fn in glob.glob('weights/*.npy'):
+    if 'moving' not in fn and 'Exponential' not in fn:
+        os.remove(fn)
+
 
 for fn in glob.glob('weights/*_RMSProp.npy'):
     os.remove(fn)
